+++ docs/meta/changelog.md
@@ -1,11 +1,16 @@
 # Changelog
 
-## Upcoming
+## 0.8
 
+### 0.8.1 - 2025-09-25 - Bugfix for frame reconstruction, CLI updates, more metadata for `stream`
+
+#### New features
 - [`#93`](https://github.com/Aharoni-Lab/mio/pull/93) - Add `mio config list` to display available configs
 - [`#103`](https://github.com/Aharoni-Lab/mio/pull/103) - Add `mio hash` to cli to hash video files
+- [`#136`](https://github.com/Aharoni-Lab/mio/pull/136) - Add runtime metadata fields for later video repair
 
-## 0.8.*
+#### Bugfixes
+- [`#136`](https://github.com/Aharoni-Lab/mio/pull/136) - Fix buffer misalignment during frame reconstruction
 
 ### 0.8.0 - 2025-09-19 - Realtime frequency filters for `stream` and test adjustments